@@ -9,11 +9,18 @@ export let setLocalStorage = (key, value)=>{
 export let getLocalStorage = (key)=>{
     const data = localStorage.getItem(key);
     return data;
-    // try{
-    //     return JSON.parse(data);
-    // }catch(err){
-    //     return err;
-    // }
 }
 
 export let removeLocalStorage = (key) => localStorage.removeItem(key);
+
+
+
+export const getCookie = (name) => {
+    const cookie = document.cookie.split('; ').find(cookie => cookie.startsWith(name + '='));
+    
+    return cookie ? cookie.substring(name.length + 1) : null;
+  };
+
+  export const getUserId = () => {
+    return getLocalStorage('user');
+  }
