@@ -4,45 +4,38 @@ import 'package:bluetooth_ecg/providers/ecg_provider.dart';
 import 'package:bluetooth_ecg/utils/files_management.dart';
 import 'package:bluetooth_ecg/utils/utils.dart';
 import 'package:dio/dio.dart';
+import 'package:http_parser/http_parser.dart';
 import 'package:provider/provider.dart';
 
 ECGProvider ecgProvider = Utils.globalContext!.read<ECGProvider>();
 
 class ECGRecordController {
   static Future<void> uploadFileToDB(Map fileUploadInformation) async {
-    final url = APIConstant.apiUrlProduction + 'ecg-records/upload';
+    final url = APIConstant.apiUrlProduction + 'api/record';
 
-    final String filePath = fileUploadInformation["filePath"];
-    final String userId = fileUploadInformation["userId"];
-    final int deviceId = fileUploadInformation["deviceId"];
-    final DateTime startTime = fileUploadInformation["startTime"];
-    final DateTime stopTime = fileUploadInformation["stopTime"];
-
+    final String filePath = fileUploadInformation["file_path"];
+    final MultipartFile fileData = await MultipartFile.fromFile(
+      filePath, 
+      filename: filePath.split('/').last,
+      contentType: MediaType('text', 'csv')
+    );
     FormData fileToUpload = FormData.fromMap({
-      'file': await MultipartFile.fromFile(filePath, filename: filePath.split('/').last),
-      'user_id': userId,
-      'device_id': deviceId,
-      'start_time': startTime,
-      'stop_time': stopTime,
-      'sensor_type': "ECG"
+      ...fileUploadInformation,
+      "file": fileData,
     });
     try {
       final response = await Dio().post(url, data: fileToUpload);
-      print('response:$response');
-    } catch (e) {
+    } catch (e, t) {
       // save filePath to preferences
       FilesManagement.saveFilePathCaseNoInternet(filePath);
-      print('error when upload file: $e');
+      print('error when upload file: $e $t');
     }
   }
 
   static Future<void> getAllECGRecords(int userId) async {
     try {
       final String url = APIConstant.apiUrlProduction + 'ecg-records/patient/$userId';
-
-      int time = DateTime.now().millisecondsSinceEpoch;
       final Response response = await Dio().get(url);
-      int real = DateTime.now().millisecondsSinceEpoch - time;
 
       final responseData = response.data;
       if (responseData["status"] == "success") {
@@ -58,9 +51,7 @@ class ECGRecordController {
     try {
       final String url = APIConstant.apiUrlProduction + 'ecg-records/record-data/$recordId';
 
-      int time = DateTime.now().millisecondsSinceEpoch;
       final Response response = await Dio().get(url);
-      int real = DateTime.now().millisecondsSinceEpoch - time;
       final responseData = response.data;
       if (responseData["status"] == "success") {
         List ecgRecordData = responseData["data"];
