@@ -20,14 +20,68 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
+  static String m0(action) => "Bluetooth is ${action} now!";
+
   final messages = _notInlinedMessages(_notInlinedMessages);
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "availableDevices":
+            MessageLookupByLibrary.simpleMessage("Available devices"),
+        "bloodPressure": MessageLookupByLibrary.simpleMessage("Blood pressure"),
+        "bluetoothStatusDes": m0,
         "chat": MessageLookupByLibrary.simpleMessage("Chat"),
+        "connect": MessageLookupByLibrary.simpleMessage("Connect"),
+        "connected": MessageLookupByLibrary.simpleMessage("Connected"),
+        "dataProcessed": MessageLookupByLibrary.simpleMessage("Data processed"),
+        "dataProcessingDes": MessageLookupByLibrary.simpleMessage(
+            "Data is being processed by Python"),
+        "devicesFoundDes":
+            MessageLookupByLibrary.simpleMessage("All devices found!"),
+        "diastolic": MessageLookupByLibrary.simpleMessage("diastolic"),
+        "disconnect": MessageLookupByLibrary.simpleMessage("Disconnect"),
+        "edit": MessageLookupByLibrary.simpleMessage("Edit"),
+        "enter": MessageLookupByLibrary.simpleMessage("Enter"),
+        "extremelyHigh": MessageLookupByLibrary.simpleMessage("Extremely high"),
+        "familyMemberPhone":
+            MessageLookupByLibrary.simpleMessage("Family member phone number"),
+        "find": MessageLookupByLibrary.simpleMessage("Find"),
+        "general": MessageLookupByLibrary.simpleMessage("General"),
+        "heartbeat": MessageLookupByLibrary.simpleMessage("Heartbeat"),
         "history": MessageLookupByLibrary.simpleMessage("History"),
         "home": MessageLookupByLibrary.simpleMessage("Home"),
+        "hypertension": MessageLookupByLibrary.simpleMessage("Hypertension"),
+        "level": MessageLookupByLibrary.simpleMessage("Level"),
         "login": MessageLookupByLibrary.simpleMessage("Login"),
+        "measure": MessageLookupByLibrary.simpleMessage("Measure"),
+        "measurementPage":
+            MessageLookupByLibrary.simpleMessage("Measurement Page"),
+        "normal": MessageLookupByLibrary.simpleMessage("Normal"),
+        "notification": MessageLookupByLibrary.simpleMessage("Notification"),
+        "or": MessageLookupByLibrary.simpleMessage("Or"),
+        "phoneNumber": MessageLookupByLibrary.simpleMessage("Phone number"),
         "profile": MessageLookupByLibrary.simpleMessage("Profile"),
         "registration": MessageLookupByLibrary.simpleMessage("Registration"),
-        "signUp": MessageLookupByLibrary.simpleMessage("Sign up")
+        "reset": MessageLookupByLibrary.simpleMessage("Reset"),
+        "result": MessageLookupByLibrary.simpleMessage("Result"),
+        "save": MessageLookupByLibrary.simpleMessage("Save"),
+        "saveAndCalculate":
+            MessageLookupByLibrary.simpleMessage("Save and calculate"),
+        "saveDataToStorage":
+            MessageLookupByLibrary.simpleMessage("Save data to storage"),
+        "scanningDes": MessageLookupByLibrary.simpleMessage(
+            "Let\'s connect your phone to hardware using Bluetooth"),
+        "scanningDevices":
+            MessageLookupByLibrary.simpleMessage("Scanning devices"),
+        "signUp": MessageLookupByLibrary.simpleMessage("Sign up"),
+        "somethingWentWrong":
+            MessageLookupByLibrary.simpleMessage("Something went wrong"),
+        "status": MessageLookupByLibrary.simpleMessage("Status"),
+        "stop": MessageLookupByLibrary.simpleMessage("Stop"),
+        "systolic": MessageLookupByLibrary.simpleMessage("Systolic"),
+        "turnOn": MessageLookupByLibrary.simpleMessage("Turn on"),
+        "variability": MessageLookupByLibrary.simpleMessage("Variability"),
+        "welcomeSentence":
+            MessageLookupByLibrary.simpleMessage("Welcome to my app,"),
+        "yourBloodPressureChart":
+            MessageLookupByLibrary.simpleMessage("Your blood pressure chart")
       };
 }
