@@ -5,9 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:permission_handler/permission_handler.dart';
 
-Uuid _UART_UUID = Uuid.parse("6E400001-B5A3-F393-E0A9-E50E24DCCA9E");
-Uuid _UART_RX   = Uuid.parse("6E400002-B5A3-F393-E0A9-E50E24DCCA9E");
-Uuid _UART_TX   = Uuid.parse("6E400003-B5A3-F393-E0A9-E50E24DCCA9E");
+Uuid uartUUID = Uuid.parse("6E400001-B5A3-F393-E0A9-E50E24DCCA9E");
+Uuid uartRX   = Uuid.parse("6E400002-B5A3-F393-E0A9-E50E24DCCA9E");
+Uuid uartTX   = Uuid.parse("6E400003-B5A3-F393-E0A9-E50E24DCCA9E");
 
 class BleReactiveScreen extends StatefulWidget {
   const BleReactiveScreen({Key? key}) : super(key: key);
@@ -121,7 +121,7 @@ class _BleReactiveScreenState extends State<BleReactiveScreen> {
     _currentConnectionStream = flutterReactiveBle.connectToAdvertisingDevice(
       id: _deviceNeedConnecting.id,
       prescanDuration: const Duration(seconds: 1),
-      withServices: [_UART_UUID, _UART_TX],
+      withServices: [uartUUID, uartTX],
     );
 
     _connection = _currentConnectionStream.listen((event) {
@@ -138,8 +138,8 @@ class _BleReactiveScreenState extends State<BleReactiveScreen> {
           });
           _numberOfMessagesReceived = 0;
           _txCharacteristic = QualifiedCharacteristic(
-            serviceId: _UART_UUID, 
-            characteristicId: _UART_TX, 
+            serviceId: uartUUID, 
+            characteristicId: uartTX, 
             deviceId: event.deviceId
           );
           // _rxCharacteristic = QualifiedCharacteristic(serviceId: _UART_UUID, characteristicId: _UART_RX, deviceId: event.deviceId);
@@ -187,7 +187,11 @@ class _BleReactiveScreenState extends State<BleReactiveScreen> {
             final state = snapshot.data;
             if (state == BleStatus.ready) {
               return const BleScanningAndConnectingScreen();
-            } else {
+            } 
+            // else if (state == BleStatus.unknown) {
+            //   return const CircularProgressIndicator();
+            // } 
+            else {
             return BluetoothOffScreen(state: state);
             }
           }),
